added base class functionality
