--- api/app/serving/config.yaml
@@ -1,30 +1,0 @@
-UniXCoder:
-  base_model: microsoft/unixcoder-base
-
-  processing:
-    batch_size: 16
-    max_length: 512
-
-  training:
-    batch_size: 128
-    max_epochs: 3
-    log_every_n_steps: 10
-    deterministic: true
-    limit_train_batches: 0.25
-    limit_val_batches: 0.25
-
-
-UniXCoder_Codesearchnet:
-  max_query_length : 128
-  max_code_length : 256
-
-  languages:  
-    python :
-      base_model_path : model_files/python/model.bin
-    javascript :
-      base_model_path : model_files/javascript/model.bin
-    go :
-      base_model_path : model_files/go/model.bin
-    java :
-      base_model_path : model_files/java/model.bin
-
