+++ api/app/serving/models/codebert.py
@@ -1,0 +1,27 @@
+import numpy as np 
+
+from .base import AbstractTransformerEncoder
+
+class CodeBert(AbstractTransformerEncoder):
+    '''
+    '''
+    def __init__(self) -> None:
+        super().__init__()
+        pass
+    
+    @property
+    def load_model(self):
+        pass
+        
+    def make_embeddings(self, config : str) -> dict:
+        '''
+        #TODO
+        '''
+        pass
+
+    def load_tokenizer(self): 
+        '''
+        '''
+        pass 
+
+
