+++ api/app/serving/models/config.yaml
@@ -1,0 +1,15 @@
+UniXCoder:
+  base_model: microsoft/unixcoder-base
+
+  processing:
+    batch_size: 16
+    max_length: 512
+
+  training:
+    batch_size: 128
+    max_epochs: 3
+    log_every_n_steps: 10
+    deterministic: true
+    limit_train_batches: 0.25
+    limit_val_batches: 0.25
+
